ai_town/requirements.txt
fastapi
uvicorn[standard]
streamlit
sentence-transformers
faiss-cpu
sqlalchemy
python-dotenv
httpx
pytest
pydantic
loguru
pdfminer.six
pypdf
